dev-requirements.txt
# For testing
django-beautifulsoup-test==1.1.1 # HTML Testing
model_mommy==1.2.6 # For creating test objects

# For development
flake8==2.5.4 # PEP8
flake8_docstrings==0.2.6 # PEP257
coverage==4.1 # Coverage checker

# For documentation
Sphinx==1.4.1

# For debugging pages
django-debug-toolbar==1.4.0
django-debug-toolbar-template-profiler==1.0.1